+++ R/use_files_external.R
@@ -14,6 +14,9 @@
 #' @param delete_zip Whether to delete the raw HTML zip after extraction. Use `"ask"` to prompt.
 #'   Only used by `use_bundled_html()` and by `use_external_html_template()`
 #'   when `url` points to a `.zip` archive.
+#' @param replace Boolean. If `TRUE`, an existing file at the target location is
+#'   overwritten. Defaults to `FALSE`, in which case the function aborts if the
+#'   target file already exists.
 #'
 #' @note See `?htmltools::htmlTemplate` and `https://shiny.rstudio.com/articles/templates.html`
 #'     for more information about `htmlTemplate`.
@@ -28,6 +31,7 @@ use_external_js_file <- function(
 	golem_wd = get_golem_wd(),
 	dir = "inst/app/www",
 	open = FALSE,
+	replace = FALSE,
 	dir_create,
 	pkg
 ) {
@@ -53,7 +57,8 @@ use_external_js_file <- function(
 		file_created_fun = after_creation_message_js,
 		golem_wd = golem_wd,
 		name = name,
-		open = open
+		open = open,
+		replace = replace
 	)
 }
 
@@ -65,6 +70,7 @@ use_external_css_file <- function(
 	golem_wd = get_golem_wd(),
 	dir = "inst/app/www",
 	open = FALSE,
+	replace = FALSE,
 	dir_create,
 	pkg
 ) {
@@ -90,7 +96,8 @@ use_external_css_file <- function(
 		file_created_fun = after_creation_message_css,
 		golem_wd = golem_wd,
 		name = name,
-		open = open
+		open = open,
+		replace = replace
 	)
 }
 
@@ -102,6 +109,7 @@ use_external_html_template <- function(
 	golem_wd = get_golem_wd(),
 	dir = "inst/app/www",
 	open = FALSE,
+	replace = FALSE,
 	dir_create,
 	extract = c("ask", "yes", "no"),
 	delete_zip = c("ask", "yes", "no")
@@ -118,6 +126,7 @@ use_external_html_template <- function(
 			golem_wd = golem_wd,
 			dir = dir,
 			open = open,
+			replace = replace,
 			extract = extract,
 			delete_zip = delete_zip
 		)
@@ -134,7 +143,8 @@ use_external_html_template <- function(
 		file_created_fun = after_creation_message_html_template,
 		golem_wd = golem_wd,
 		name = name,
-		open = open
+		open = open,
+		replace = replace
 	)
 }
 
@@ -150,6 +160,7 @@ use_external_file <- function(
 	golem_wd = get_golem_wd(),
 	dir = "inst/app/www",
 	open = FALSE,
+	replace = FALSE,
 	dir_create,
 	pkg
 ) {
@@ -175,7 +186,8 @@ use_external_file <- function(
 		file_created_fun = NULL,
 		golem_wd = golem_wd,
 		name = name,
-		open = open
+		open = open,
+		replace = replace
 	)
 }
 
@@ -187,6 +199,7 @@ use_bundled_html <- function(
 	golem_wd = get_golem_wd(),
 	dir = "inst/app/www",
 	open = FALSE,
+	replace = FALSE,
 	extract = c("ask", "yes", "no"),
 	delete_zip = c("ask", "yes", "no")
 ) {
@@ -210,18 +223,28 @@ use_bundled_html <- function(
 			sprintf("%s_bundle.zip", name_bundle)
 		)
 	}
-	check_file_exists(where_zip)
+	check_file_exists(where_zip, replace = replace)
 	check_directory_exists(dir)
 
 	where_bundle <- fs_path(dir, name_bundle)
-	if (fs_dir_exists(where_bundle) || fs_file_exists(where_bundle)) {
+	if (
+		!isTRUE(replace) &&
+			(fs_dir_exists(where_bundle) || fs_file_exists(where_bundle))
+	) {
 		cli_abort(
 			sprintf(
-				"%s already exists.\n\nYou can delete it with:\nunlink('%s', recursive = TRUE).",
+				"%s already exists.\n\nYou can delete it with:\nunlink('%s', recursive = TRUE).\n\nAlternatively, call this function again with `replace = TRUE`.",
 				where_bundle,
 				where_bundle
 			)
 		)
+	}
+	if (isTRUE(replace)) {
+		if (fs_dir_exists(where_bundle)) {
+			fs_dir_delete(where_bundle)
+		} else if (fs_file_exists(where_bundle)) {
+			fs_file_delete(where_bundle)
+		}
 	}
 
 	progress_id <- cli_progress_bar(