+++ R/use_files_external_tools.R
@@ -94,6 +94,7 @@ perform_checks_and_download_if_everything_is_ok <- function(
 	golem_wd,
 	name,
 	open,
+	replace = FALSE,
 	pkg
 ) {
 	signal_arg_is_deprecated(
@@ -144,7 +145,8 @@ perform_checks_and_download_if_everything_is_ok <- function(
 		directory_to_download_to
 	)
 	check_file_exists(
-		where_to_download_to
+		where_to_download_to,
+		replace = replace
 	)
 	download_external(
 		url_to_download_from = url_to_download_from,